Utility to link apps and vis to lighttpd
This makes some assumptions on the name of the apps and visualizations
as registered in MongoDB and the name of the directory inside the
repository.

The name is slugified to match the name. For this purpose the directory
for the Text Record List visualizations was renamed to
text-record-list which is its slugified version.

This introduces a dependency on the awesome-slugify library
for the script.